Component refactoring | Adding contexts git commit -m
